Cleaning up more files
