+++ login/handlers.go
@@ -11,6 +11,8 @@ import (
 	"github.com/Hucaru/Valhalla/channel/player"
 	"github.com/Hucaru/Valhalla/common"
 	"github.com/Hucaru/Valhalla/common/opcode"
+	"github.com/Hucaru/Valhalla/constant"
+	"github.com/Hucaru/Valhalla/internal"
 	"github.com/Hucaru/Valhalla/mnet"
 	"github.com/Hucaru/Valhalla/mpacket"
 )
@@ -40,6 +42,7 @@ func (server *Server) HandleClientPacket(conn mnet.Client, reader mpacket.Reader
 		log.Println("UNKNOWN CLIENT PACKET:", reader)
 	}
 }
+
 func (server *Server) handleLoginRequest(conn mnet.Client, reader mpacket.Reader) {
 	username := reader.ReadString(reader.ReadInt16())
 	password := reader.ReadString(reader.ReadInt16())
@@ -373,4 +376,94 @@ func (server *Server) addCharacterItem(characterID int64, itemID int32, slot int
 
 func (server *Server) handleReturnToLoginScreen(conn mnet.Client, reader mpacket.Reader) {
 	conn.Send(packetLoginReturnFromChannel())
+}
+
+// HandleServerPacket from world
+func (server *Server) HandleServerPacket(conn mnet.Server, reader mpacket.Reader) {
+	switch reader.ReadByte() {
+	case opcode.WorldNew:
+		server.handleNewWorld(conn, reader)
+	case opcode.WorldInfo:
+		server.handleWorldInfo(conn, reader)
+	default:
+		log.Println("UNKNOWN WORLD PACKET:", reader)
+	}
+}
+
+// The following logic could do with being cleaned up
+func (server *Server) handleNewWorld(conn mnet.Server, reader mpacket.Reader) {
+	log.Println("Server register request from", conn)
+	if len(server.worlds) > 14 {
+		log.Println("Rejected")
+		conn.Send(mpacket.CreateInternal(opcode.WorldRequestBad))
+	} else {
+		name := reader.ReadString(reader.ReadInt16())
+
+		if name == "" {
+			name = constant.WORLD_NAMES[len(server.worlds)]
+
+			registered := false
+			for i, v := range server.worlds {
+				if v.Conn == nil {
+					server.worlds[i].Conn = conn
+					name = server.worlds[i].Name
+
+					registered = true
+					break
+				}
+			}
+
+			if !registered {
+				server.worlds = append(server.worlds, internal.World{Conn: conn, Name: name})
+			}
+
+			p := mpacket.CreateInternal(opcode.WorldRequestOk)
+			p.WriteString(name)
+			conn.Send(p)
+
+			log.Println("Registered", name)
+		} else {
+			registered := false
+			for i, w := range server.worlds {
+				if w.Name == name {
+					server.worlds[i].Conn = conn
+					server.worlds[i].Name = name
+
+					p := mpacket.CreateInternal(opcode.WorldRequestOk)
+					p.WriteString(name)
+					conn.Send(p)
+
+					registered = true
+
+					break
+				}
+			}
+
+			if !registered {
+				server.worlds = append(server.worlds, internal.World{Conn: conn, Name: name})
+
+				p := mpacket.CreateInternal(opcode.WorldRequestOk)
+				p.WriteString(server.worlds[len(server.worlds)-1].Name)
+				conn.Send(p)
+			}
+
+			log.Println("Re-registered", name)
+		}
+	}
+}
+
+func (server *Server) handleWorldInfo(conn mnet.Server, reader mpacket.Reader) {
+	for i, v := range server.worlds {
+		if v.Conn != conn {
+			continue
+		}
+
+		server.worlds[i].SerialisePacket(reader)
+
+		if v.Name == "" {
+			log.Println("Registerd new world", server.worlds[i].Name)
+		} else {
+			log.Println("Updated world info for", v.Name)
+		}
+	}
 }
